requirements.txt
requests
python-dotenv
openai
llama_index
pymilvus